replace with threshold parameter
